[chore] 테스트용 api 컨트롤러 분리
